Object-Oriented Programming
Mini game that consisted in a grid with a fireman that was controlled by the user. Fire spread randomly, with the odds being related to the type of terrain attached next to it (Eucaliptus, pine, dry grass). The fireman could put out fires, or with the press of a key, summon an airplane which doused water on the column with the most active fires.