Object-Oriented Programming
FireFighter Simulator 
The project consisted in a game where the user controlled a fireman dousing fires. 
The fires spread randomly, and depending on the type of terrain. For instances, the probability of fire spreading on dry grass is higher than on pinewood. 
The user can embark on a bulldozer, to turn patches of terrain to dry land and with the push of the "C" button, a plane is summoned that douses the column with the most active fires.
Project done on the year 2017.